refactor: share the YouTube id scan instead of triplicating it
The Sonar gate failed on duplication — 22.1% against a 3% threshold —
and it was mine. Per-file figures name it exactly: 78 new lines in each
of the three video modules, 100% / 100% / 71.8% duplicated. Nothing else
in the PR duplicates, so this one block is the whole gate failure.

`packages/survey-ui/src/lib/youtube-id.ts` now owns the scan and the
three modules import it. survey-ui is the only workable home: it is the
leaf of the three, `surveys` already depends on it, and `apps/web` now
does too.

No new build entry was needed. survey-ui builds with
`preserveModules: true`, so `dist/lib/youtube-id.{js,d.ts}` falls out of
the existing config and only the `./youtube-id` export map entry had to
be added.

The three video modules stay separate otherwise — they are parallel by
design, one per build target, and were already about half-identical
before this PR. Only the scan is shared, because its greedy,
backtracking and line-terminator semantics are subtle enough that three
copies would be three chances to get them wrong.

Behaviour is unchanged: still identical to base across 50k inputs after
the extraction, all three suites green, catalog check passes, and the
web dependency graph builds with the new cross-package import.
